fix: Measure items properly in single page viewmode
